Move AlchemyEncoder to jsonencoders
